html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-tokenization-width-vs-innerwidth.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML: window.open `features`: tokenization -- size features `width` vs `innerwidth`</title>
<meta name=timeout content=long>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#apis-for-creating-and-navigating-browsing-contexts-by-name">

<!-- user agents are not required to support open features other than `noopener`
     and on some platforms position and size features don't make sense -->
<meta name="flags" content="may" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/PrefixedPostMessage.js"></script>
<script>
var windowURL = 'resources/message-opener.html';
var featuresPrefix = `top=0,left=0,height=401,`;

setup (() => {
  // Before running tests, open a window using features that mimic
  // what would happen if the feature tested here were set to 0
  // for comparison later.
  var featureString = `${featuresPrefix}width=0`;
  var prefixedMessage = new PrefixedMessageTest();
  prefixedMessage.onMessage((data, e) => {
    e.source.close();
    runWindowTests(data);
  });
  var win = window.open(prefixedMessage.url(windowURL), '', featureString);
});

function runWindowTests (baselineDimensions) {

  // The last specified feature is used
  [ 'width=401,width=402',
  [ 'width=error,width=402',
    'width=401,innerwidth=402',
    'width=error,innerwidth=402',
    'innerwidth=401,width=402',
    'innerwidth=error,width=402',
    'innerwidth=401,innerwidth=402',
    'innerwidth=error,innerwidth=402',
  ].forEach(features => {
    async_test(t => {
      var prefixedMessage = new PrefixedMessageTest();
      prefixedMessage.onMessage(t.step_func_done((data, e) => {
        e.source.close();
        assert_equals(data.width, 402);
      }));
      var win = window.open(prefixedMessage.url(windowURL), '', featuresPrefix + features);
    }, `${format_value(features)} repeated width/innerwidth features`);
  });

  // The last specified feature is used even if the value is 0 or error
  [ 'width=401,width=0',
  [ 'width=error,width=0',
    'width=401,innerwidth=0',
    'width=error,innerwidth=0',
    'innerwidth=401,width=0',
    'innerwidth=error,width=0',
    'innerwidth=401,innerwidth=0',
    'innerwidth=error,innerwidth=0',
    'width=401,width=error',
    'width=error,width=error',
    'width=401,innerwidth=error',
    'width=error,innerwidth=error',
    'innerwidth=401,width=error',
    'innerwidth=error,width=error',
    'innerwidth=401,innerwidth=error',
    'innerwidth=error,innerwidth=error',
  ].forEach(features => {
    async_test(t => {
      var prefixedMessage = new PrefixedMessageTest();
      prefixedMessage.onMessage(t.step_func_done((data, e) => {
        e.source.close();
        assert_equals(data.width, baselineDimensions.width);
      }));
      var win = window.open(prefixedMessage.url(windowURL), '', featuresPrefix + features);
    }, `${format_value(features)} repeated width/innerwidth features where the last one is 0 or error`);
  });
}

</script>
